refactor(commands): Component structure
